fix : todo 뷰, 로직 관심사 분리
